// packages/core/src/native/registry/Registry.types.ts
export interface Coin {
  amount: Uint128
  denom: string
}
// ...
export type Namespace = string
export type ModuleVersion =
  | 'latest'
  | {
      version: string
    }
export type ModuleReference =
  | {
      account_base: number
    }
  | {
      native: Addr
    }
  | {
      adapter: Addr
    }
  | {
      app: number
    }
  | {
      standalone: number
    }
export type Addr = string
export type UpdateModule =
  | {
      default: {
        metadata: string
      }
    }
  | {
      versioned: {
        instantiation_funds?: Coin[] | null
        metadata?: string | null
        monetization?: Monetization | null
        version: string
      }
    }
export type Monetization =
  | 'none'
  | {
      install_fee: FixedFee
    }
// ...
export interface ModuleInfo {
  name: string
  namespace: Namespace
  version: ModuleVersion
}
// ...
export type ModuleStatus =
  | 'r_e_g_i_s_t_e_r_e_d'
  | 'p_e_n_d_i_n_g'
  | 'y_a_n_k_e_d'
export interface ModuleFilter {
  name?: string | null
  namespace?: string | null
  status?: ModuleStatus | null
  version?: string | null
}
export type MigrateMsg = {}
export interface AccountBaseResponse {
  account_base: AccountBase
}
export interface ConfigResponse {
  factory: Addr
}
export interface ModulesListResponse {
  modules: ModuleResponse[]
}
export interface ModuleResponse {
  config: ModuleConfiguration
  module: Module
}
export interface ModuleConfiguration {
  instantiation_funds: Coin[]
  metadata?: string | null
  monetization: Monetization
}
export interface Module {
  info: ModuleInfo
  reference: ModuleReference
}